+++ requirements.txt
@@ -1,2 +1,3 @@
 flask
 pyjwt
+flask_sqlalchemy
